refactor(InputFile): parts.tsx/style.tsをclient/へ移動し循環依存を解消する
InputFile.tsxが依存するのはtypes・共有hooks・client/のみで、parts.tsx/style.tsは
実質client専用のファイルだった。client/index.tsがDownloadAnchorButton/InputFileNative/
InputFileMultiplyAppendableを同一バレルでre-exportしていたため、rollupビルドで
client/index.ts→InputFileNative.tsx→parts.tsx→client/index.tsという循環依存が
実際に発生していた（parts.tsxがDownloadAnchorButtonをバレル経由でimportしていたため）。

parts.tsx/style.tsをclient/直下に移動し、DownloadAnchorButtonへの参照をバレル経由から
同一ディレクトリの直接参照に変更することで循環依存を解消した。client/index.tsからは
不要になったDownloadAnchorButtonのexportを削除し、types.tsが型のみ必要としていた
classNameGeneratorをexport typeで追加した。

rollupビルドで循環依存が解消したこと、sandbox/nextのnext buildで
RSCCheckerが引き続き"server component"と判定することを確認済み。
