+++ packages/smarthr-ui/src/components/InputFile/client/InputFileNative.tsx
@@ -14,10 +14,10 @@ import { useMergeRefs } from '../../../hooks/client/useMergeRefs'
 import { useLatest } from '../../../hooks/useLatest'
 import { Stack } from '../../Layout'
 import { Groupbox } from '../../Panel'
-import { FileListItem, LabelRender, StyledFaFolderOpenIcon } from '../parts'
-import { classNameGenerator } from '../style'
 
 import { FilePreviewDialog } from './FilePreviewDialog'
+import { FileListItem, LabelRender, StyledFaFolderOpenIcon } from './parts'
+import { classNameGenerator } from './style'
 
 import type { LowerProps } from '../types'
