+++ packages/smarthr-ui/src/components/InputFile/client/style.ts
